Fix lot of silly errors.
o We're in the FDLIBM package

kernel-cos:
o For kernel-cos, ix is the top part with the signed removed.
o The c coefficients need to be doubles, not floats!
o The test for qx was wrong.
o Use ix and not x to make the value of qx.

kernel-sin:
o ix is the top part with the signed removed.
o Fix computation of the result: was not including y in the right
  places.

kernel-tan:
o Need to KERNEL package prefix in many places.